# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopEasy e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  offerPrice: number;
  /** @nullable */
  discountPercent?: number | null;
  images: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  inStock: boolean;
  featured?: boolean;
  /** @nullable */
  stockCount?: number | null;
  createdAt: string;
}