# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopEasy e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  offerPrice: number;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  inStock?: boolean;
  featured?: boolean;
  /** @nullable */
  stockCount?: number | null;
}