# lib/api-zod/src/generated/types/stockUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopEasy e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface StockUpdate {
  inStock: boolean;
  /** @nullable */
  stockCount?: number | null;
}